# Speed up some tests