Trying to identify the tests that take the longest and make them faster.  So far working on `test_grid.py`.  Previously this took ~15min on my machine:
```
========================================================================================== 25 passed, 2 warnings in 993.05s (0:16:33) ==========================================================================================
pytest tests/test_grid.py  368.17s user 130.13s system 50% cpu 16:36.10 total
```
but now it takes 5 (just made the grids smaller):
```
========================================================================================== 25 passed, 2 warnings in 332.38s (0:05:32) ==========================================================================================
pytest tests/test_grid.py  197.18s user 61.64s system 77% cpu 5:35.38 total
```