@dlakaplan 2023-10-21
Slowest 20 remaining tests:
```
============================= slowest 20 durations =============================
228.56s call     tests/test_tim_writing.py::test_flags
112.02s call     tests/test_toa_indexing.py::test_getitem_boolean
111.43s call     tests/test_toa_indexing.py::test_select
111.29s call     tests/test_toa_indexing.py::test_getitem_where
110.43s call     tests/test_toa_indexing.py::test_getitem_slice
48.87s call     tests/test_fake_toas.py::test_fake_DMfit
42.52s call     tests/test_toa_reader.py::test_load_multiple
42.46s call     tests/test_toa_reader.py::test_pickle_multiple
37.84s call     tests/test_grid.py::test_grid_3param_prefix_singleprocessor
36.54s call     tests/test_polycos.py::test_generate_polycos[72-9-400.0-gbt]
35.31s call     tests/test_grid.py::test_grid_3param_singleprocessor
34.52s call     tests/test_polycos.py::test_generate_polycos[72-9-1400.0-gbt]
34.51s call     tests/test_polycos.py::test_generate_polycos[72-9-400.0-coe]
34.29s call     tests/test_polycos.py::test_generate_polycos[72-9-400.0-ao]
34.16s call     tests/test_polycos.py::test_generate_polycos[72-9-1400.0-ao]
33.98s call     tests/test_polycos.py::test_generate_polycos[72-9-1400.0-coe]
26.25s call     tests/test_grid.py::test_grid_3param_prefix_multiprocessor
25.64s call     tests/test_toa_reader.py::test_renumber_subset_reordered
25.04s call     tests/test_toa_reader.py::test_merge_indices
24.99s call     tests/test_toa_reader.py::test_merge_indices_excised
```