@dlakaplan 2023-10-21
Many of those are slow because they are going from tim files to toas and back for many permutations.  Individual tests are fast but there are many of them.  I'm wary of changing them,