@dlakaplan 2023-10-23
I think that this can still be worked on more generally, but as a minor PR this saves ~10 minutes from the CI testing and can probably be merged as-is.